refactor: 네이밍 수정 및 디렉토리 삭제
