+++ src/main/java/domain/lotto/LottoNumberGenerator.java
@@ -1,10 +1,10 @@
-package domain.lotto.number;
+package domain.lotto;
 
 import domain.lotto.generator.NumberGeneratable;
 
 import java.util.List;
 
-public class LottoNumber {
+public class LottoNumberGenerator {
     public List<Integer> makeNumbers(NumberGeneratable numberGeneratable) {
         return numberGeneratable.generate();
     }
